backend/tests/test_main.py
"""
Basic tests for OpenShelf backend API.

Uses pytest-asyncio and httpx AsyncClient with mocked database dependencies.
"""

from unittest.mock import AsyncMock, MagicMock, patch

import pytest
import pytest_asyncio
from httpx import ASGITransport, AsyncClient


# ---------------------------------------------------------------------------
# App import — gracefully skip if dependencies aren't installed yet
# ---------------------------------------------------------------------------
try:
    from main import app
except ImportError:
    import sys
    from pathlib import Path

    sys.path.insert(0, str(Path(__file__).parent.parent))
    try:
        from main import app
    except Exception:
        app = None


# ---------------------------------------------------------------------------
# Fixtures
# ---------------------------------------------------------------------------


@pytest_asyncio.fixture
async def client():
    """Async HTTP client wired to the FastAPI app."""
    if app is None:
        pytest.skip("Could not import app — skipping integration tests")
    async with AsyncClient(
        transport=ASGITransport(app=app), base_url="http://test"
    ) as ac:
        yield ac


# ---------------------------------------------------------------------------
# Health check
# ---------------------------------------------------------------------------


@pytest.mark.asyncio
async def test_health_check(client):
    """GET / should return 200 with {"status": "ok"}."""
    response = await client.get("/")
    assert response.status_code == 200
    data = response.json()
    assert data.get("status") == "ok"


# ---------------------------------------------------------------------------
# Books endpoint
# ---------------------------------------------------------------------------


@pytest.mark.asyncio
async def test_get_books_empty(client):
    """GET /books should return 200 with a paginated result containing an items list."""
    # Patch the database session to return an empty result
    mock_result = MagicMock()
    mock_result.scalars.return_value.all.return_value = []
    mock_result.scalar.return_value = 0

    mock_session = AsyncMock()
    mock_session.execute.return_value = mock_result

    with patch("main.get_db", return_value=mock_session):
        response = await client.get("/books")

    assert response.status_code == 200
    data = response.json()
    # Accept either {"items": [...]} or a list directly
    if isinstance(data, dict):
        assert "items" in data
        assert isinstance(data["items"], list)
    else:
        assert isinstance(data, list)


# ---------------------------------------------------------------------------
# Stats endpoint
# ---------------------------------------------------------------------------


@pytest.mark.asyncio
async def test_get_stats(client):
    """GET /stats should return 200 with expected keys."""
    mock_result = MagicMock()
    mock_result.scalar.return_value = 0

    mock_session = AsyncMock()
    mock_session.execute.return_value = mock_result

    with patch("main.get_db", return_value=mock_session):
        response = await client.get("/stats")

    assert response.status_code == 200
    data = response.json()
    assert isinstance(data, dict)

    # At minimum we expect some counting keys
    expected_keys = {"total_books", "total_authors", "languages"}
    # Accept partial match — at least one key must be present
    found = expected_keys.intersection(data.keys())
    assert found, (
        f"Expected at least one of {expected_keys} in response, got: {list(data.keys())}"
    )


# ---------------------------------------------------------------------------
# Sanity — no-app fallback (runs even without the real app)
# ---------------------------------------------------------------------------


def test_placeholder_always_passes():
    """Placeholder test that always passes — confirms pytest is wired correctly."""
    assert 1 + 1 == 2
