backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
azure-storage-blob==12.19.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
httpx==0.27.0
aiofiles==23.2.1
python-slugify==8.0.4
chardet==5.2.0
EbookLib==0.18
